Remove unneeded depedency to the Database Cleaner gem
I noticed you were the Database Cleaner gem but I don't think it's necessary anymore since in recent versions of Rails every test is already run in a transaction, even for feature tests.

I remvoved the gem, ran the tests, and everything is still working so I guess we're good!