Remove duplicate pending migrations checks for testing
I noticed there were 2 calls for this same check (3 if you count the one in [`components/courses/spec/rails_helper.rb`][3], but it doesn't seem you're using it): 1) in this [`rails_helper`][1] and 2) in the [`spec_helper`][2], so we can remove one of them.

Since `rails_helper` requires `spec_helper` in the beginning, it seems safe to remove the later one in `rails_helper`.

[1]: https://github.com/pivorakmeetup/pivorak-web-app/blob/d55f825fe7c5ac062fa51411c84b614519d026f6/spec/rails_helper.rb#L29
[2]: https://github.com/pivorakmeetup/pivorak-web-app/blob/d55f825fe7c5ac062fa51411c84b614519d026f6/spec/spec_helper.rb#L40
[3]: https://github.com/pivorakmeetup/pivorak-web-app/blob/d55f825fe7c5ac062fa51411c84b614519d026f6/components/courses/spec/rails_helper.rb#L19